Refactor scan simulator within scan device to give more options for subclasses.
